Transform now creates an array of arguments for each event.
Bean takes arguments in several different forms and many of them are optional.
This makes it difficult to transform the arguments properly and it has taken a
while to arrive at a simple solution.

This commit fixes any abmiguity regarding optional argument position and
honestly I'm surprised it every worked before.