Fixed an bug in afterAnimation
Bug: If an animation delays its start, the cuttoff may be reached and the
callback is triggered. Then if the animation starts later, it will
trigger a second callback.

Fix: When triggering the callback at the cutoff, remove the event
listener for animation start, so callback won't be fired twice.

Consideration: Setting a longer checkStart delay is recommended if
animation timing is causing callbacks to be fired early.